+++ backend/src/main/java/com/preproject/backend/domain/member/service/MemberService.java
@@ -3,7 +3,7 @@ package com.preproject.backend.domain.member.service;
 import com.preproject.backend.domain.member.entity.Member;
 import com.preproject.backend.domain.member.repository.MemberRepository;
 import com.preproject.backend.global.auth.utils.CustomAuthorityUtils;
-import com.preproject.backend.global.helper.event.MemberRegistrationApplicationEvent;
+import com.preproject.backend.global.auth.utils.GetAuthUserUtils;
 import com.preproject.backend.global.exception.BusinessLogicException;
 import com.preproject.backend.global.exception.ExceptionCode;
 import com.preproject.backend.global.utils.CustomBeanUtils;
@@ -11,19 +11,15 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Isolation;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
 
-@Transactional
 @Service
+@Transactional
 @RequiredArgsConstructor
 public class MemberService {
     private final MemberRepository memberRepository;
@@ -31,7 +27,7 @@ public class MemberService {
     private final PasswordEncoder passwordEncoder;
     private final CustomAuthorityUtils authorityUtils;
 
-    // TODO member 등록 (회원가입) / 보안 적용
+    // member 등록 (회원가입)
     public Member createMember(Member member) {
         verifyExistsEmail(member.getEmail());
 
@@ -47,11 +43,14 @@ public class MemberService {
     }
 
     // member 수정
-    @Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.SERIALIZABLE)
+    //@Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.SERIALIZABLE)
     public Member updateMember(Member member) {
         verifyExistsEmail2(member.getEmail()); // 먼저 해당 멤버의 이메일이 존재하는지 확인
 
         Member findMember = findVerifiedMember(member.getMemberId()); // 존재한다면 해당 아이디의 멤버 가져와서
+
+        if(getLoginMember().getMemberId() != member.getMemberId()) // 로그인 유저 != 작성자 이면
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_MEMBER);
 
         beanUtils.copyNonNullProperties(member, findMember); // 수정
 
@@ -75,17 +74,20 @@ public class MemberService {
     public void deleteMember(long memberId) {
         Member findMember = findVerifiedMember(memberId);
 
+        if(getLoginMember().getMemberId() != findMember.getMemberId()) // 로그인 유저 != 작성자 이면
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_MEMBER);
+
         memberRepository.delete(findMember);
     }
 
     // member id로 검색
-    @Transactional(readOnly = true)
+    //@Transactional(readOnly = true)
     public Member findVerifiedMember(long memberId) {
         Optional<Member> optionalMember =
                 memberRepository.findById(memberId);
 
         return optionalMember.orElseThrow(() -> {
-                    return new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND);
+                    throw  new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND);
                 });
     }
 
@@ -100,24 +102,14 @@ public class MemberService {
     // member email 이 존재하는지 검증 --> 수정 시, 같은 email 이여야지만 수정이 가능하도록 ( 해당 email 이 없다면 exception 반환 )
     private void verifyExistsEmail2(String email) {
         Optional<Member> member = memberRepository.findByEmail(email);
-
         if (member.isEmpty())
             throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND);
     }
 
     // Login 한 Member 를 가져오는 로직
     public Member getLoginMember() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        // 여기 "anonymousUser" 부분은 security 코드를 보고 권한이 없는 Member 의 이름이 어떻게 되어있는지 확인 후 수정해야할 것 같습니다 !
-        if(authentication == null || authentication.getName() == null || authentication.getName().equals("anonymousMember"))
-            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_MEMBER);
-
-        Optional<Member> optionalUser = memberRepository.findByEmail(authentication.getName());
-        Member member = optionalUser.orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
-
-        System.out.println("memberId : " + member.getMemberId());
-
-        return member;
+        return  memberRepository.findByEmail(GetAuthUserUtils.getAuthUser().getName())
+                .orElseThrow(()
+                        -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));
     }
 }
